+++ CHANGELOG.md
@@ -12,7 +12,7 @@
 
 ### New Features
 
-_None_
+- Adds `if_exists` parameter to `upload_to_s3` action, with possible values `:skip`, `:fail`, and `:replace`. [#495]
 
 ### Bug Fixes